chore: cleanup deployment root files package
Helper methods on the deployment root package were moved
to the appropriate subpackages to improve code organization and
ownership.
