+++ deployment/keystone/changeset/shared/labeled_addresses_test.go
@@ -1,4 +1,4 @@
-package deployment
+package shared
 
 import (
 	"testing"
